moved to sane types (f.e.: uint16_t -> u16)

+++ src/idt/interrupts.c
@@ -1,3 +1,4 @@
+#include "util/types.h"
 #include "interrupts.h"
 
 #define SCREEN_X 80
@@ -6,12 +7,12 @@
 static char* screen = (char*)0xb8000;
 
 __attribute__((no_caller_saved_registers)) void screen_clear() {
-    for (int i = 0; i < SCREEN_X * SCREEN_Y * 2; i++) {
+    for (u16 i = 0; i < SCREEN_X * SCREEN_Y * 2; i++) {
         screen[i] = i;
     }
 }
 __attribute__((no_caller_saved_registers)) void screen_print(char* text) {
-    for (int i = 0; *text != '\0'; i += 2, text++) {
+    for (u16 i = 0; *text != '\0'; i += 2, text++) {
         screen[i] = *text;
         screen[i + 1] = 4;
     }
